Alteração de execução do script, adição das variáveis de classe nos Rs e Rq, correção nos controllers

+++ src/main/java/br/com/projetob2w/springbootcommysql/repository/CargoRepository.java
@@ -10,6 +10,6 @@ import javax.persistence.NamedNativeQuery;
 import java.util.List;
 
 @Repository
-public interface CargoRepository extends JpaRepository<Cargo, Long> {
+public interface CargoRepository extends JpaRepository<Cargo, Integer> {
     List<Cargo> findByCargoNomeContains(String name);
 }
